Other: Rename tables to table.
