+++ src/tableui.js
@@ -4,10 +4,10 @@
  */
 
 /**
- * @module tables/tablesui
+ * @module table/tableui
  */
 
-import Plugin from '../../ckeditor5-core/src/plugin';
+import Plugin from '@ckeditor/ckeditor5-core/src/plugin';
 
-export default class TablesUI extends Plugin {
+export default class TableUI extends Plugin {
 }
