+++ tests/tableediting.js
@@ -3,20 +3,20 @@
  * For licensing, see LICENSE.md.
  */
 
-import TablesEditing from '../src/tablesediting';
+import TableEditing from '../src/tableediting';
 
-import Paragraph from '../../ckeditor5-paragraph/src/paragraph';
+import Paragraph from '@ckeditor/ckeditor5-paragraph/src/paragraph';
 
-import VirtualTestEditor from '../../ckeditor5-core/tests/_utils/virtualtesteditor';
-import { getData as getModelData, setData as setModelData } from '../../ckeditor5-engine/src/dev-utils/model';
+import VirtualTestEditor from '@ckeditor/ckeditor5-core/tests/_utils/virtualtesteditor';
+import { getData as getModelData, setData as setModelData } from '@ckeditor/ckeditor5-engine/src/dev-utils/model';
 
-describe( 'TablesEditing', () => {
+describe( 'TableEditing', () => {
 	let editor, model;
 
 	beforeEach( () => {
 		return VirtualTestEditor
 			.create( {
-				plugins: [ TablesEditing, Paragraph ]
+				plugins: [ TableEditing, Paragraph ]
 			} )
 			.then( newEditor => {
 				editor = newEditor;
